fix: correct missing 'paths' element in before_compile context options
If `cordova prepare` and `cordova compile` are run independently, the paths array generated by prepare is not passed on to compile, and therefore not available to the `before_compile` script. This is normally not an issue if the user runs `cordova build` which runs prepare and compile in a promise chain with the same options object.

closes issue #122