fix: set proper module include path and check it before loading
Module include paths used by `require` are not file-system relative, and if your file is not in one of the known locations the path must be absolute in order to prevent "Cannot find module" errors. This commit ensures that the path used for loading `Api.js` is absolute based on the current working directory, and add a check to ensure that the file exists before trying to require it.