fix: replicate cordova prepare functionality for getting platform paths (#153)
* fix: replicate cordova prepare functionality for getting platform paths

Blindly using platforms/<platform>/www is not correct, as each platform can have the www folder in a different location. A more robust approach uses the platforms functionality of cordova-lib to obtain the path to the www folder.

* fix: remove external cordova-lib dependency

* fix: use os-safe path separators

* fix: set proper module include path and check it before loading

Module include paths used by `require` are not file-system relative, and if your file is not in one of the known locations the path must be absolute in order to prevent "Cannot find module" errors. This commit ensures that the path used for loading `Api.js` is absolute based on the current working directory, and add a check to ensure that the file exists before trying to require it.

* revert: undo whitespace disaster

This reverts commit cde3007c70875d419268afecbe968f730feaa13d.

* fix: re-commit without whitespace changes