fix(ios): Handle auto session tracking start on iOS (#210)
Same as getsentry/sentry-react-native#1308

Tested on iOS Simulator with sample app. With this, release health now works on the Cordova SDK for both iOS and Android. Browser will need a unique solution to use @sentry/browser's sessions while disabling it when running on native platforms. This will be added in a later version.