fix: Remove setRelease and setDist, have auto release passed to native
As the Android native bridge has no support for setRelease and setDist which set __sentry_release and __sentry_dist as an extra, we remove these methods in favor of passing release and dist to init.
We also have the SENTRY_RELEASE environment variable that is set at buildtime set the release on init instead.

Closes #196