fix: Remove setRelease and setDist, have auto release passed to native  (#213)
In previous versions, we hadsetRelease and setDist which set __sentry_release and __sentry_dist as an extra, we remove these methods in favor of passing release and dist to init. This removes the existing code on the iOS native bridge that handles this, and we did not add support for this on the Android bridge.

We also have the SENTRY_RELEASE environment variable that is built by the script at build time and matches the sourcemaps set the release on init instead of setting it in a JS event processor. This way we have the auto release set on native events as well.

Closes #196